+++ configs/all-js.js
@@ -4,10 +4,10 @@ import {languageOptions} from './base.js';
 
 /* eslint-disable @stylistic/no-multi-spaces */
 import getConfigJs   from '../plugin-configs/@eslint/js/all.js';
-import getConfigNode from '../plugin-configs/eslint-plugin-/node/all.js';
+import getConfigNode from '../plugin-configs/eslint-plugin-n/all.js';
 
 // Import getConfigImport  from '../plugin-configs/eslint-plugin-/import/all.js';
-import getConfigPromise from '../plugin-configs/eslint-plugin-/promise/all.js';
+import getConfigPromise from '../plugin-configs/eslint-plugin-promise/all.js';
 
 import getConfigEslintComments from '../plugin-configs/@eslint-community/eslint-plugin-eslint-comments/format.js';
 import getConfigStylistic      from '../plugin-configs/@stylistic/format.js';
