+++ jsconfig.json
@@ -4,6 +4,7 @@
 		"strict": true,
 		"esModuleInterop": true,
 		"noImplicitAny": true,
+		"noEmit": true,
 		"module": "ESNext",
 		"target": "ESNext",
 		"moduleResolution": "bundler"
