+++ plugin-configs/eslint-plugin-import/rules/deprecated.js
