+++ plugin-configs/eslint-plugin-import/rules/static-analysis.js
