+++ plugin-configs/eslint-plugin-n/all.js
