+++ plugin-configs/eslint-plugin-n/rules/deprecated.js
