+++ plugin-configs/eslint-plugin-n/rules/prefer-global.js
