+++ plugin-configs/eslint-plugin-n/rules/prefer-promises.js
