+++ plugin-configs/eslint-plugin-n/rules/without-recommended.js
