+++ plugin-configs/eslint-plugin-promise/all.js
