+++ plugin-configs/eslint-plugin-promise/rules/all.js
